+++ src/lib.rs
@@ -186,13 +186,13 @@ pub use comment::{
 pub use encoder::{
     pack_residue_classification_groups, pack_residue_classifications, residue_body_shape,
     residue_partition_codeword_count, write_audio_packet_header, write_codebook,
-    write_comment_header, write_floor0_header, write_floor1_header, write_floor1_packet,
-    write_identification_header, write_mapping_header, write_mode_header, write_residue_body,
-    write_residue_header, write_residue_partition, Floor1Packet, PackResidueClassError,
-    PackResidueClassGroupsError, ResidueBodyShape, ResidueVectorPlan, WriteAudioPacketHeaderError,
-    WriteCodebookError, WriteError, WriteFloor0Error, WriteFloor1Error, WriteFloor1PacketError,
-    WriteMappingError, WriteModeError, WriteResidueBodyError, WriteResidueError,
-    WriteResiduePartitionError,
+    write_comment_header, write_floor0_header, write_floor0_packet, write_floor1_header,
+    write_floor1_packet, write_identification_header, write_mapping_header, write_mode_header,
+    write_residue_body, write_residue_header, write_residue_partition, Floor0Packet, Floor1Packet,
+    PackResidueClassError, PackResidueClassGroupsError, ResidueBodyShape, ResidueVectorPlan,
+    WriteAudioPacketHeaderError, WriteCodebookError, WriteError, WriteFloor0Error,
+    WriteFloor0PacketError, WriteFloor1Error, WriteFloor1PacketError, WriteMappingError,
+    WriteModeError, WriteResidueBodyError, WriteResidueError, WriteResiduePartitionError,
 };
 pub use floor0::{bark as floor0_bark, Floor0Curve, Floor0Decoder, Floor0Error};
 pub use floor1::{